clean up app, delete unneccessary html, more renaming
