+++ src/server/templates/index.html
@@ -37,18 +37,6 @@
     <div class="row mt-2">
 
         <div class="form-group">
-            <div class="form-row">
-                <div class="col">
-                    <input class="form-control" id="new_x" name="new_x" placeholder="relative x"/>
-                </div>
-                <div class="col">
-                    <input class="form-control" id="new_y" name="new_y" placeholder="relative y"/>
-                </div>
-                <div class="col">
-                    <button id="change" class="btn btn-primary btn-block" onclick="change()">change relative coords
-                    </button>
-                </div>
-            </div>
             <div id="battery">
                 Battery: <p id="level"></p>
             </div>
