+++ tests/AlgorithmTest.py
@@ -1,5 +1,5 @@
 import math
-from src.movement import FollowingDrone
+from src.Movement import FollowingDrone
 import time
 from threading import Thread
